+++ src/httprpc.cpp
@@ -4,7 +4,6 @@
 
 #include <httprpc.h>
 
-#include <chainparams.h>
 #include <common/args.h>
 #include <config.h>
 #include <crypto/hmac_sha256.h>
@@ -12,7 +11,6 @@
 #include <rpc/protocol.h>
 #include <util/strencodings.h>
 #include <util/string.h>
-#include <util/translation.h>
 #include <walletinitinterface.h>
 
 #include <boost/algorithm/string.hpp>
@@ -25,6 +23,7 @@
 #include <memory>
 #include <set>
 #include <string>
+#include <vector>
 
 /** WWW-Authenticate to present with 401 Unauthorized response */
 static const char *WWW_AUTH_HEADER_DATA = "Basic realm=\"jsonrpc\"";
@@ -72,6 +71,8 @@ static std::string strRPCUserColonPass;
 static std::string strRPCCORSDomain;
 /* Stored RPC timer interface (for unregistration) */
 static std::unique_ptr<HTTPRPCTimerInterface> httpRPCTimerInterface;
+/* List of -rpcauth values */
+static std::vector<std::vector<std::string>> g_rpcauth;
 /* RPC Auth Whitelist */
 static std::map<std::string, std::set<std::string>> g_rpc_whitelist;
 static bool g_rpc_whitelist_default = false;
@@ -105,15 +106,7 @@ static bool multiUserAuthorized(std::string strUserPass) {
     std::string strUser = strUserPass.substr(0, strUserPass.find(':'));
     std::string strPass = strUserPass.substr(strUserPass.find(':') + 1);
 
-    for (const std::string &strRPCAuth : gArgs.GetArgs("-rpcauth")) {
-        // Search for multi-user login/pass "rpcauth" from config
-        std::vector<std::string> vFields;
-        boost::split(vFields, strRPCAuth, boost::is_any_of(":$"));
-        if (vFields.size() != 3) {
-            // Incorrect formatting in config file
-            continue;
-        }
-
+    for (const auto &vFields : g_rpcauth) {
         std::string strName = vFields[0];
         if (!TimingResistantEqual(strName, strUser)) {
             continue;
@@ -425,6 +418,19 @@ static bool InitRPCAuthentication() {
 
     if (gArgs.GetArg("-rpcauth", "") != "") {
         LogPrintf("Using rpcauth authentication.\n");
+        for (const std::string &rpcauth : gArgs.GetArgs("-rpcauth")) {
+            std::vector<std::string> fields{SplitString(rpcauth, ':')};
+            const std::vector<std::string> salt_hmac{
+                SplitString(fields.back(), '$')};
+            if (fields.size() == 2 && salt_hmac.size() == 2) {
+                fields.pop_back();
+                fields.insert(fields.end(), salt_hmac.begin(), salt_hmac.end());
+                g_rpcauth.push_back(fields);
+            } else {
+                LogPrintf("Invalid -rpcauth argument.\n");
+                return false;
+            }
+        }
     }
 
     g_rpc_whitelist_default = gArgs.GetBoolArg("-rpcwhitelistdefault",